Move more lane transformation logic into raw_map
